Relocate shared SMB transport under network/smb/common (#519)
Move network/smb/transport to network/smb/common/transport to group
version-agnostic SMB building blocks under a common/ namespace, and
repoint the smb_v10 client and dcerpc doc-comment references. Pure
relocation, no behavior change.
