+++ network/smb/common/transport/transport.go
