+++ network/smb/common/transport/transport_test.go
@@ -4,7 +4,7 @@ import (
 	"net"
 	"testing"
 
-	"github.com/TheManticoreProject/Manticore/network/smb/transport"
+	"github.com/TheManticoreProject/Manticore/network/smb/common/transport"
 )
 
 // MockTransport implements the Transport interface for testing
